[BUGFIX] Rack::Locale fails to detect properly formatted locale names
The convention is that locales use uppercase for the country specific
portion of the name (e.g. "en-AU"), see:

https://github.com/svenfuchs/rails-i18n

The author also used this as a chance to switch to using mutating array
operations, like `[Array#map!]` over `[Array#map]` to avoid allocating
so many objects in a critical path.